extract interface and make this public instead of Impl
